+++ frontend/js/utils.js
@@ -15,7 +15,6 @@ export function throttle(callback, delay) {
             lastThis = this;
             return;
         }
-        }
 
         callback.apply(this, args);
         isThrottled = true;